# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Password Vault API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  username: string;
  email: string;
  password: string;
  fullName: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  fullName: string;
  role: string;
  createdAt: string;
  updatedAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface UserSummary {
  id: number;
  username: string;
  email: string;
  fullName: string;
  role: string;
}

export interface UpdateUserBody {
  email?: string;
  fullName?: string;
  role?: string;
}

export interface VaultEntry {
  key: string;
  value: string;
}

export interface VaultItemSummary {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description: string | null;
  accessControl: string;
  entryCount: number;
  createdBy: number;
  createdByUsername: string;
  createdAt: string;
  updatedAt: string;
}

export interface VaultItem {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description: string | null;
  accessControl: string;
  allowedUserIds: number[];
  entries: VaultEntry[];
  createdBy: number;
  createdByUsername: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateVaultItemBody {
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  accessControl: string;
  allowedUserIds?: number[];
  entries: VaultEntry[];
}

export interface UpdateVaultItemBody {
  name?: string;
  category?: string;
  /** @nullable */
  description?: string | null;
  accessControl?: string;
  allowedUserIds?: number[];
  entries?: VaultEntry[];
}

export interface VaultStats {
  totalItems: number;
  credentialCount: number;
  globalVarCount: number;
  totalEntries: number;
  accessibleToMe: number;
}

export interface ApiKey {
  id: number;
  name: string;
  keyPrefix: string;
  isActive: boolean;
  /** @nullable */
  lastUsedAt: string | null;
  createdAt: string;
}

export interface CreateApiKeyBody {
  name: string;
}

export interface ApiKeyCreated {
  id: number;
  name: string;
  keyPrefix: string;
  rawKey: string;
  createdAt: string;
}

export interface Certificate {
  id: number;
  name: string;
  fingerprint: string;
  isActive: boolean;
  expiresAt: string;
  createdAt: string;
}

export interface GenerateCertificateBody {
  name: string;
  validityDays: number;
}

export interface CertificateCreated {
  id: number;
  name: string;
  fingerprint: string;
  isActive: boolean;
  expiresAt: string;
  createdAt: string;
  pemBundle: string;
}

export interface CertificateDownload {
  id: number;
  name: string;
  pemBundle: string;
  fingerprint: string;
}

export interface AuditLog {
  id: number;
  userId: number;
  username: string;
  vaultItemId: number;
  vaultItemName: string;
  action: string;
  /** @nullable */
  ipAddress: string | null;
  /** @nullable */
  userAgent: string | null;
  createdAt: string;
}

export interface AuditLogPage {
  logs: AuditLog[];
  total: number;
  page: number;
  pageSize: number;
}

export interface TopAccessedItem {
  vaultItemId: number;
  vaultItemName: string;
  accessCount: number;
}

export interface AuditStats {
  totalAccesses: number;
  uniqueUsers: number;
  topAccessedItems: TopAccessedItem[];
  recentActivity: AuditLog[];
}

export type ListAuditLogsParams = {
  /**
   * @nullable
   */
  vaultItemId?: number | null;
  /**
   * @nullable
   */
  userId?: number | null;
  page?: number;
  pageSize?: number;
};