# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Password Vault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Vault Items
...
  /apikeys:
    get:
      operationId: listApiKeys
      tags: [apikeys]
      summary: List API keys for current user
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"

    post:
      operationId: createApiKey
      tags: [apikeys]
      summary: Generate a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyBody"
      responses:
        "201":
          description: Created API key (raw value shown only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
  /apikeys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [apikeys]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked

  # Certificates
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates for current user
      responses:
        "200":
          description: List of certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"

    post:
      operationId: generateCertificate
      tags: [certificates]
      summary: Generate a new client certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCertificateBody"
      responses:
        "201":
          description: Certificate generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateCreated"
  /certificates/{id}:
    delete:
      operationId: revokeCertificate
      tags: [certificates]
      summary: Revoke a certificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked
  /certificates/{id}/download:
    get:
      operationId: downloadCertificate
      tags: [certificates]
      summary: Download certificate bundle (PEM) — web UI only
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Certificate PEM file download info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateDownload"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Audit Logs
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth schemas
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
      required:
        - username
        - email
        - password
        - fullName
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword

    # User schemas
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - username
        - email
        - fullName
        - role
        - createdAt
        - updatedAt
...
    UpdateUserBody:
      type: object
      properties:
        email:
          type: string
        fullName:
          type: string
        role:
          type: string

    # Vault schemas
...
    ApiKey:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        isActive:
          type: boolean
        lastUsedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - keyPrefix
        - isActive
        - lastUsedAt
        - createdAt
    CreateApiKeyBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    ApiKeyCreated:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        rawKey:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - keyPrefix
        - rawKey
        - createdAt

    # Certificate schemas
    Certificate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fingerprint:
          type: string
        isActive:
          type: boolean
        expiresAt:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - fingerprint
        - isActive
        - expiresAt
        - createdAt
    GenerateCertificateBody:
      type: object
      properties:
        name:
          type: string
        validityDays:
          type: integer
      required:
        - name
        - validityDays
    CertificateCreated:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fingerprint:
          type: string
        isActive:
          type: boolean
        expiresAt:
          type: string
        createdAt:
          type: string
        pemBundle:
          type: string
      required:
        - id
        - name
        - fingerprint
        - isActive
        - expiresAt
        - createdAt
        - pemBundle
    CertificateDownload:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        pemBundle:
          type: string
        fingerprint:
          type: string
      required:
        - id
        - name
        - pemBundle
        - fingerprint

    # Audit Log schemas
...